refactor(comparison): extract projection model
